Update sasLogParserMacros.sas
Added logic (lines 12-15) based on operating system to set a new macro variable &delm to / for linux and \ for windows.

Modified line 25 to use the new macro variable &delm. which allows the utility to do a recursive parsing of log files in the parent directory as well as all sub-directories.